Refactor message handling to use MESSAGES_CONFIG_MANAGER and introduce CacheConfig for global boost data management
