+++ common/src/main/java/dev/matthiesen/common/cobblemon_boosters/interfaces/IWebhookService.java
@@ -1,7 +1,7 @@
 package dev.matthiesen.common.cobblemon_boosters.interfaces;
 
-import dev.matthiesen.common.cobblemon_boosters.config.ModConfig;
+import dev.matthiesen.common.cobblemon_boosters.config.WebhooksConfig;
 
 public interface IWebhookService {
-    void sendMessage(ModConfig.DiscordEmbed embed, IBoost boost);
+    void sendMessage(WebhooksConfig.DiscordEmbed embed, IBoost boost);
 }
