+++ common/src/main/java/dev/matthiesen/common/cobblemon_boosters/utils/NoOpWebhookService.java
@@ -1,12 +1,12 @@
 package dev.matthiesen.common.cobblemon_boosters.utils;
 
-import dev.matthiesen.common.cobblemon_boosters.config.ModConfig;
+import dev.matthiesen.common.cobblemon_boosters.config.WebhooksConfig;
 import dev.matthiesen.common.cobblemon_boosters.interfaces.IBoost;
 import dev.matthiesen.common.cobblemon_boosters.interfaces.IWebhookService;
 
 public class NoOpWebhookService implements IWebhookService {
     @Override
-    public void sendMessage(ModConfig.DiscordEmbed embed, IBoost boost) {
+    public void sendMessage(WebhooksConfig.DiscordEmbed embed, IBoost boost) {
         // No operation performed, as this is a no-op implementation.
     }
 }
